Add disable_general_search configuration option to prevent searches
As part of my investigation in to issue #8, I've added the configuration
option disable_general_search (Boolean) to the config schema. Now, when
search is called (and it isn't a uri based search), it checked whether
disable_general_search is set to true. If it is true, then it parses the
search parameters to see if there is a valid URL with youtube.com in it.
If it finds youtube.com, the search will recurse but this time, the uri
will be set in the query so that it performs a search on the URL passed
in.

This should mean that the extension only goes to Youtube if it's
definitely a URL for youtube.