Add support for high quality YouTube Music streams for Premium users. (#221)
This requires to supply yt-dlp with a cookiefile that exists on disk.
The `musicapi_cookie` config option is replaced with a
`musicapi_cookiefile` option. This also allows the cookies to be better
protected by giving the file proper permissions or using Docker secrets
in a Docker deployment.

Some nasty glue code is needed to generate a Cookie header from the
CookieJar (There is no functionality in the Python stdlib to do this
cleanly). A better approach would be to refactor all code to take a
CookieJar opbject (including ytmusicapi) that can be natively used in
requests and urllib (assuming it is used somewhere). As this requires
some work on a dependency I am refraining from doing this in this PR.

Co-authored-by: natumbri <natumbri@gmail.com>